+++ conda_verify/checks.py
@@ -10,7 +10,7 @@ from conda_verify.common import (check_build_number, check_build_string,
                                  get_python_version_specs)
 from conda_verify.constants import FIELDS, LICENSE_FAMILIES
 from conda_verify.exceptions import PackageError, RecipeError
-from conda_verify.utils import all_ascii, get_bad_seq, get_field, get_object_type
+from conda_verify.utilities import all_ascii, get_bad_seq, get_field, get_object_type
 
 
 class CondaPackageCheck(object):
