+++ conda_verify/utilities.py
